requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
networkx
torch
gymnasium
pyyaml
python-dotenv
requests
matplotlib
scipy
websockets
streamlit
plotly
requests

